Make BaseEntity.equals stricter.
Previously, even objects of different types, like a Game and a
GameTitle, could be equal. Now, two objects are equal if and only if
either:

	1. They are identical objects.
	2. Both refer to the same database entity.

Importantly, if two objects refer to the same entity, but they do not
have identical properties, they are still equal. The tests included
should clarify the exact behavior of the method.